+++ mail/mail-uniffi/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
   - [ET-1999] `Message::is_draft` property.
+  - `EventError` now reports refresh error.
 
 ### Fixed